Refactored utilities from LIL into a mixin class to use with other sparse matrix classes.
